feat(EMS-2433): code and test fixes
