+++ src/ui/server/controllers/insurance/eligibility/buyer-country/index.ts
@@ -93,7 +93,7 @@ export const post = async (req: Request, res: Response) => {
 
     /**
      * If a country has no insurance support and no short term cover,
-     * redirect to contract too short page.
+     * redirect to CONTRACT_TOO_SHORT.
      */
     const noShortTermCover = !country.noInsuranceSupport && !country.shortTermCover;
